l4virtio server: fix size type of registered dataspaces
The register_ds() function takes a size parameter of l4_umword_t type
which denotes the size of the dataspace. Unfortunately the size gets
truncated to 32 bits while passing it to _check_n_init() which is called
from register_ds() because the size parameter of _check_n_init() is only
a l4_uint32_t. This makes it impossible for virtio guests to register a
dataspace larger than 4GB.

This change changes all 'size' parameters to be of l4_umword_t type.
This allows clients on 64 bit architectures to register dataspaces
larger than 4GB. It also adds an overflow check to the contains()
function.

Change-Id: I037274ff1736d8ad69733be3764a6d41c398adce