virtio-block: publish device features after setting them
virtio-block host implementations are able to set their host features.
The features are stored in a local bitmap but they are never published
into the config space where a virtio guest can read the features from.

This change makes sure that we publish updated host features into the
config space upon reset of the virtio device.

Change-Id: I1d435c2c799b436bd0886b9894fbb5cec1a78bd7