Data_buffer: allow copying a part of one buffer to another
copy_to() always copies the maximum number of bytes possible, but there
is no way to copy a chosen maximum number of bytes smaller than that.
Add a default parameter to copy_to() which allows specifying a maximum
number of bytes, that copies either at most that much or what fits.  The
parameter defaults to UINT_MAX, thus providing the old behaviour if not
specified.

Change-Id: I571eb3152981f835ae3bcdf3deade8905ff67035