Detect and reject legacy drivers by checking feature map.
A non-transitional device must offer VIRTIO_F_VERSION_1 and may choose
to fail to operate if it is not accepted. We opt to fail in this case.

Legacy interface support is optional and the L4virtio implementation
does not support it so cause the internal feature check to fail when
L4VIRTIO_FEATURE_VERSION_1 isn't set in the driver feature map. This
change thus prevents the FEATURES_OK state from being reached.

Change-Id: I444aa1914401d57563fb62f5f3c3f18ef7bc3fef