fix memory barrier for certain x86 32-bit archs
We should not test for __i686__ to detect if memory barrier functions
are available. For instance, for -m32 -march=k8, __i686__ is not defined
but rather __i386__ and __k8__.

It's probably better to assume that 'sfence' and 'lfence' exist these
days even on 32-bit x86 processors so test also for __i386__ which is
always present.

Change-Id: I6f1f98d845e2340341950bc40097f78dcc741b36