Unify virtio-block client implementations
In this commit we also remove support for implicit driver state
reinitialization in the device because it was incompatible with the
unified client implementation. In order to reinitialize, servers must
provide other means (such as an explicit RPC) that resets the driver and
reinitializes its state in the device.

Fixes #CD-285.
Fixes #L4RE-353.

Change-Id: If2452aeeb2296ee4cbdbf3b8541d75e1621382e3