virtio_input: Fix input event value size
The value of Virtio_input_event is currently defined as 16-bit number,
but according to the VIRTIO specification it is actually supposed to be
a signed 32-bit little-endian integer.

This makes it impossible to send input events that are larger than
16-bit, e.g. ABS_MT_TRACKING_ID = -1 = 0xffffffff (for a touchscreen) is
currently always truncated to 0xffff = 65535 before reaching the client.

Change-Id: Ie13db762e82688fbfc1a55b783b5cd95547a0943