Implement trusted dataspace validation
Applications using the l4virtio library now have the option to
validate the dataspaces for queues and buffers as trusted ones.

A virtio device now holds the functions
`enable_trusted_ds_validation()` and `add_trusted_dataspaces()` to
enable the trusted dataspace validation and to add a set of trusted
dataspaces for the validation.

A dataspace for queues and buffers provided by a client, is a
trusted one, if it is included in the list of trusted dataspaces.

This change also adjusts the default virtio-block client to accept
an external queue dataspace as an option to allocating the queue
dataspace itself.

Change-Id: I2660da51ca734ae4807e85b24c3cc4347ad4fdb0