fix opcodes to provide compatibility with older clients
Change I7101ffee3d0053f32c71860ee2bbb89b6172c9b0 removed the
register_iface() method from the API. This caused all opcodes to change
as well. This may break older clients which are not compiled against
the newer library.

To fix this we manually define the opcodes in an enum and assign them
via the appropriate RPC framework macros. That way we can assure
compatiblity between older and newer clients and hosts.

Change-Id: I538b37ece8df4ac78b44a55a3f463c002b0ae0ce