fix + optimize Bad_descriptor::message()
Make the array static. Otherwise the compiler will generate the array
on the stack if the array index is not known. With a static array, also
check if the array element is not nullptr. An element could contain a
nullptr if the developer forgot to initialize an array element in the
middle of the array.

Also fix the style by moving the opening brace to the next line.

Change-Id: I7bbd1f81aa7ace5b3c037b589a4ce43b49ae0edc