virtqueue: optimize barriers on armv8+
Virtio works on shared, cache coherent memory. Thus, it is sufficient
to order the memory accesses with respect to each other. They do not
necessarily need to be retired because all potential observers are
supposed to be in the same inner shareability domain. This means that
we can consistently use DMBs on Arm, like Armv7 did already previously.

Change-Id: I6e024e3e75c5adbfeaf0e0fd7f5615308fa72ae1