virtio-console-device: remove broken ctrl queue retries
The L4virtio::Svr::Console::Device class tried to re-send control queue
messages based on the port state if this failed previously.
Unfortunately, the old code did not account for the fact that multiple
messages might be necessary. As this is now handled transparently in
the base class, the affectively dead code can be removed.

Change-Id: I1db6e53d96fb79c1504249382517a77db2b491ec