virtio-console: correctly handle exhausted ctrl queue
The device must report its ports and their state to the driver through
the control queue. IF the control queue RX descriptors are exhausted,
the message must be sent once new buffers are available.

Previously, this duty was delegated to the super class of
L4virtio::Svr::Console::Virtio_con. This is not necessary because all
information is available already and the control queue is actually
dispatched in the base class too.

Implementation wise, each port gains a "reported state" field. When the
control queue is exhausted, the reported state will lack behind the
actual port state. Depending on the previously reported state and the
current state, the messages are sent once new buffers are available. If
the message could be sent, the reported state is advanced.

Change-Id: I397eaf7201751c401ec2a6f15dc37276bcbf47a3