virtio-console: fix process_port_open() description
The VIRTIO_CONSOLE_PORT_OPEN message can be sent by both the driver and
the device. Regarding the device, this is not an acknowledgement as the
comment suggested previously but the port state on the device side.
The driver message is thus just an indication of the port state on the
driver side. The device does not need to act upon it.

Change-Id: I23169dfdbaa5c6318b4821cadc8628b69e3a0447