[API] client/virtio-net: Use IRQ instead of semaphore to notify driver
The semaphore approach is based on the assumption that the device sends
a notification for each used item. But that is not necessarily the
case. When the device consecutively processes multiple items, not to
be confused with chained buffers, it might opt to just send a single
notification at the end. Thus, the semaphore count can get out of
sync with the number of items used by the device, resulting in the
client blocking on the semaphore's down operation even though used
items are available in the virtqueue.

Without the aforementioned assumption, a semaphore is unsuitable for
use as a driver notification. But even ignoring that, it is questionable
for the client to make a Semaphore::down() syscall for each used buffer.

Therefore, replace the semaphore with an IRQ, and before waiting for a
driver notification in wait_rx(), first check whether used
items are available in the virtqueue.

The downside is that other than a Semaphore, the IRQ must be bound to
the waiting thread beforehand.

Change-Id: I91ed27b1dab9bea665d58c1543380a52e123aa56