l4virtio: set device notification IRQ index
A device may implement multiple device notification IRQs. The driver
needs to be told which interrupt shall be used for the device
configuration or a particular virtqueue. This is done by the
device_notify_index fields in the config header or the queue
configuration. Add ways to set those fields.

Change-Id: Ie9c8948b95c5afbb519d89c77e60f09a63109b53